Move isEOA to utils
