+++ tests/quote.spec.ts
@@ -1,12 +1,11 @@
 import { handleQuoteRequest } from "@/src/app/api/tools/quote/logic";
 import { withRedactedErrorHandling } from "@/src/lib/error";
-import { isEOA } from "@/src/lib/protocol/quote";
-import { EthRpc, ParsedQuoteRequest } from "@/src/lib/types";
-import { getClientForChain } from "@bitte-ai/agent-sdk";
+import { ParsedQuoteRequest } from "@/src/lib/types";
 import { OrderQuoteRequest } from "@cowprotocol/cow-sdk";
+
 const MAX_VALID_FROM = 10800; // 3 hours
-describe("Quote Route Logic", () => {
-  it.skip("should transform parsedQuoteRequest into Full Quote", async () => {
+describe.skip("Quote Route Logic", () => {
+  it("should transform parsedQuoteRequest into Full Quote", async () => {
     const quoteRequest = {
       sellToken: "0xEeeeeEeeeEeEeeEeEeEeeEEEeeeeEeeeeeeeEEeE", // ETH
       buyToken: "0x833589fCD6eDb6E08f4c7C32D4f71b54bdA02913",
@@ -44,11 +43,4 @@ describe("Quote Route Logic", () => {
       withRedactedErrorHandling(handleQuoteRequest(input)),
     ).rejects.toThrow("ExcessiveValidTo: validTo is too far into the future");
   }, 10000);
-
-  it.only("should transform parsedQuoteRequest into Full Quote", async () => {
-    const client = getClientForChain(8453) as EthRpc;
-    expect(
-      await isEOA(client, "0x66268791B55e1F5fA585D990326519F101407257"),
-    ).toBe(true);
-  });
 });
